vector now returns int values

+++ includes/ft_vector.h
@@ -6,7 +6,7 @@
 /*   By: user <user@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/04/03 12:29:23 by user              #+#    #+#             */
-/*   Updated: 2020/04/24 16:05:58 by user             ###   ########.fr       */
+/*   Updated: 2020/04/26 21:50:02 by user             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@
 # include "ft_memory.h"
 # include "mode_libft.h"
 
-void		*vec_create(size_t elems, size_t type_size);
+void		*vec_create(int elems, int type_size);
 void		vec_destroy(t_vector vecptr);
 
 void		vec_pushback(t_vector vecptr, void *elem);
@@ -27,10 +27,10 @@ void		vec_popback(t_vector vector);
 
 void		*vec_copy(t_vector vecptr, bool save_capacity);
 void		vec_shrink_to_fit(t_vector vecptr);
-void		vec_reserve(t_vector vecptr, size_t size);
+void		vec_reserve(t_vector vecptr, int size);
 
-size_t		vec_size(t_vector vecptr);
-size_t		vec_capacity(t_vector vecptr);
-size_t		vec_typesize(t_vector vecptr);
+int			vec_size(t_vector vecptr);
+int			vec_capacity(t_vector vecptr);
+int			vec_typesize(t_vector vecptr);
 
 #endif
