+++ vector/vec_create.c
@@ -6,7 +6,7 @@
 /*   By: user <user@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/03/06 19:31:44 by dbendu            #+#    #+#             */
-/*   Updated: 2020/03/28 10:09:44 by user             ###   ########.fr       */
+/*   Updated: 2020/04/26 21:50:39 by user             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@
 
 # ifdef SAFE_MODE
 
-void			*vec_create(size_t elems, size_t type_size)
+void			*vec_create(int elems, int type_size)
 {
 	t_vector_s	*vector;
